A couple streaming fixes.
1. Don't set 'foo' when 'foo__stream' is accessed. Previously that set
action was happening and it was leading to weirdness like the inability
to call 'dumps()' due to the partial objects in 'foo' and also
overwriting any values that were added to 'foo' before 'foo__stream' was
accessed.

2. If 'foo' was accessed before 'foo_stream', wrap the contents of 'foo'
in callables and return them. No real streaming happens, but it respects
the interface and returns the latest data for 'foo' instead of whatever
was in the buffer.